Correct the sort order
I found when running this action the sort order wasn't correct. By adding a sort field to each item as it's added it keeps things consistent.

Furthermore, when attempting to load the repeater field as a draft (via ProDrafts) I got an error. This was fixed by adding a matrx_sort param to each matrix item of a repeater_matrix.